feat(web): Add CedarApolloProvider. Deprecate RedwoodApolloProvider
