+++ packages/web/src/components/FetchConfigProvider.tsx
@@ -24,8 +24,9 @@ interface Props {
 /**
  * The `FetchConfigProvider` understands Redwood's Auth and determines the
  * correct request-headers based on a user's authentication state.
- * Note that the auth bearer token is now passed in packages/web/src/apollo/index.tsx
- * as the token is retrieved async
+ * Note that the auth bearer token is now passed in
+ * packages/web/src/apollo/ApolloProviderWithFetchConfig.tsx as the token is
+ * retrieved async
  */
 export const FetchConfigProvider: React.FC<Props> = ({
   useAuth = useNoAuth,
